Add/Remove custom conditions from combatant
